rails db:migrateを実行

+++ db/schema.rb
@@ -35,7 +35,7 @@ ActiveRecord::Schema.define(version: 2020_12_27_052521) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.bigint "blob_id", null: false
+    t.integer "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
